Feature: Login과정 수정 및 추가정보 입력 과정 수정
