+++ backend/tos-proejct/src/main/java/com/togethersports/tosproejct/user/UserController.java
@@ -1,12 +1,13 @@
 package com.togethersports.tosproejct.user;
 
-import com.togethersports.tosproejct.user.dto.*;
-import com.togethersports.tosproejct.user.exception.NicknameDuplicationException;
 import com.togethersports.tosproejct.common.code.CommonCode;
 import com.togethersports.tosproejct.common.dto.Response;
 import com.togethersports.tosproejct.security.annotation.CurrentUser;
+import com.togethersports.tosproejct.user.dto.UserOfModifyInfo;
+import com.togethersports.tosproejct.user.dto.UserOfMyInfo;
+import com.togethersports.tosproejct.user.dto.UserOfOtherInfo;
+import com.togethersports.tosproejct.user.exception.NicknameDuplicationException;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -20,21 +21,12 @@ import org.springframework.web.bind.annotation.*;
  * @author seunjeon
  * @author younghoCha
  */
-@Slf4j
+
 @RequiredArgsConstructor
 @RestController
 public class UserController {
 
     private final UserService userService;
-
-    // 계정 최초 로그인 시 추가정보 입력 처리
-    @PostMapping("/api/user")
-    public ResponseEntity<Response<?>> updateUserInfo(@CurrentUser User user,
-                                                      @RequestBody @Validated UserOfInitInfo userOfInfoUpdate) {
-        userService.initUserInfo(user.getId(), userOfInfoUpdate);
-
-        return ResponseEntity.ok().body(Response.of(CommonCode.GOOD_REQUEST, null));
-    }
 
     // 닉네임 중복확인
     @GetMapping("/api/duplication")
@@ -57,7 +49,7 @@ public class UserController {
 
     }
     //내 정보 수정
-    @PutMapping("/api/user")
+    @PostMapping("/api/user")
     public ResponseEntity<Response> modifyMyInfo(@CurrentUser User user,
                                                  @RequestBody @Validated UserOfModifyInfo userOfOtherInfo){
