+++ backend/tos-proejct/src/main/java/com/togethersports/tosproejct/user/dto/UserOfModifyInfo.java
@@ -3,6 +3,8 @@ package com.togethersports.tosproejct.user.dto;
 import com.togethersports.tosproejct.user.Gender;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
@@ -24,13 +26,21 @@ public class UserOfModifyInfo {
     @NotNull(message = "닉네임 값이 누락되었습니다.")
     private String userNickname;
 
-    private String userProfileImage;
+    private UserProfileImage userProfileImage;
 
     @NotNull(message = "생년월일이 누락되었습니다.")
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
     private LocalDate userBirth;
 
     @NotNull(message = "성별이 누락되었습니다.")
     private Gender gender;
+
+    @Setter
+    @Getter
+    public static class UserProfileImage {
+        private String userProfileExtension;
+        private String imageSource;
+    }
 
 
 }
